Address PR feedback
Simplify the HTTP CONNECT section by bringing the example closer
to the explanation of this functionality and removing some
ambiguity over whether one or both of HTTPS_PROXY and HTTP_PROXY are
required. The original text suggested that either is required, then
that both are required. The new text says to assign both, then
explains why.